feat: 사용자 채팅 기록 코드 구현 및 역할에 따른 서비스 분리
