+++ src/main/java/com/project/baedalsodae/recommendation/entity/enums/MessageType.java
@@ -1,0 +1,7 @@
+package com.project.baedalsodae.recommendation.entity.enums;
+
+public enum MessageType {
+    USER,
+    ASSISTANT,
+    SYSTEM
+}
