+++ src/main/java/com/project/baedalsodae/recommendation/service/RecommendationService.java
@@ -1,92 +1,46 @@
 package com.project.baedalsodae.recommendation.service;
 
-import com.project.baedalsodae.menu.entity.MenuItem;
-import com.project.baedalsodae.menu.repository.MenuItemRepository;
 import com.project.baedalsodae.recommendation.dto.VoiceRecommendationRequest;
 import com.project.baedalsodae.recommendation.dto.VoiceRecommendationResponse;
 import com.project.baedalsodae.recommendation.dto.VoiceRecommendationResponse.RecommendedMenu;
-import com.project.baedalsodae.store.entity.Store;
-import com.project.baedalsodae.store.repository.StoreRepository;
-import java.util.ArrayList;
-import java.util.Comparator;
 import java.util.List;
-import java.util.UUID;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.ai.chat.client.ChatClient;
-import org.springframework.ai.document.Document;
-import org.springframework.ai.vectorstore.SearchRequest;
-import org.springframework.ai.vectorstore.VectorStore;
+import org.springframework.ai.chat.memory.ChatMemory;
 import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
 
 @Slf4j
 @Service
 @RequiredArgsConstructor
 public class RecommendationService {
 
-    private final VectorStore vectorStore;
-    private final MenuItemRepository menuItemRepository;
-    private final StoreRepository storeRepository;
     private final ChatClient chatClient;
+    private final MenuSearchService menuSearchService;
 
-    @Transactional(readOnly = true)
     public VoiceRecommendationResponse recommendMenuItems(VoiceRecommendationRequest request) {
-        // 1. Vector Search
-        List<Document> similarDocuments =
-                vectorStore.similaritySearch(
-                        SearchRequest.query(request.transcribedText()).withTopK(10));
+        // 1. 읽기 트랜잭션 (빠르게 종료)
+        List<RecommendedMenu> top3 =
+                menuSearchService.recommendMenuItems(request.transcribedText());
 
-        if (similarDocuments.isEmpty()) {
+        if (top3.isEmpty()) {
             return new VoiceRecommendationResponse("죄송해요, 원하시는 조건에 맞는 메뉴를 찾지 못했어요.", List.of());
         }
 
-        // 2. Fetch Entities and Apply Weighted Scoring
-        List<RecommendedMenu> recommendations = new ArrayList<>();
-
-        record MenuScore(MenuItem menuItem, Store store, double score) {}
-        List<MenuScore> scoredMenus = new ArrayList<>();
-
-        for (Document doc : similarDocuments) {
-            String menuIdStr = (String) doc.getMetadata().get("menuId");
-            if (menuIdStr == null) continue;
-
-            UUID menuId = UUID.fromString(menuIdStr);
-            menuItemRepository
-                    .findByIdAndDeletedIsFalse(menuId)
-                    .ifPresent(
-                            menuItem -> {
-                                Store store = menuItem.getMenuCategory().getStore();
-                                double score = store.getAvgRating() * 2.0;
-                                if (menuItem.isPopular()) score += 1.0;
-
-                                scoredMenus.add(new MenuScore(menuItem, store, score));
-                            });
-        }
-
-        // 3. Sort by score descending and take Top 3
-        List<RecommendedMenu> top3 =
-                scoredMenus.stream()
-                        .sorted(Comparator.comparingDouble(MenuScore::score).reversed())
-                        .limit(3)
-                        .map(
-                                ms ->
-                                        new RecommendedMenu(
-                                                ms.store().getId(),
-                                                ms.store().getName(),
-                                                ms.store().getAvgRating(),
-                                                ms.menuItem().getId(),
-                                                ms.menuItem().getName(),
-                                                ms.menuItem().getPrice().intValue(),
-                                                ms.menuItem().getDescription()))
-                        .toList();
-
-        // 4. Generate AI Message using ChatClient
-        String prompt = buildPrompt(request.transcribedText(), top3);
+        // 2. AI 호출 + ChatMemory 저장 (트랜잭션 없음)
         String aiMessage = "추천할 메뉴를 찾았어요!";
-
         try {
-            aiMessage = chatClient.prompt().user(prompt).call().content();
+            aiMessage =
+                    chatClient
+                            .prompt()
+                            .user(buildPrompt(request.transcribedText(), top3))
+                            .advisors(
+                                    a ->
+                                            a.param(
+                                                    ChatMemory.CONVERSATION_ID,
+                                                    request.userId().toString()))
+                            .call()
+                            .content();
         } catch (Exception e) {
             log.error("Failed to generate AI message", e);
         }
@@ -109,7 +63,8 @@ public class RecommendationService {
                     .append(m.rating())
                     .append(")\n");
         }
-        sb.append("\n위의 내용을 바탕으로 사용자에게 친절하게 음성으로 읽어줄 1~2문장의 자연스러운 추천 멘트를 작성해 주세요.");
+        sb.append("\n위의 내용을 바탕으로 사용자에게 친절하게 음성으로 읽어줄 1~2문장의 자연스러운 추천 멘트를 작성해 주세요.\n");
+        sb.append("[출력 조건] 추천 멘트는 **100자**를 넘을 수 없습니다. ");
         return sb.toString();
     }
 }
